Added channel settings and moved routing in extra file
