+++ Intectainment/webpages/templates/main/channel/channelSettings.html
@@ -12,13 +12,37 @@
 
     <table>
         <tr>
-            <td>Name</td>
-            <td>{{channel.name}}</td>
-            <td>{{channel.categories}}</td>
+            <td>Kategorien</td>
+            <td>
+                <table>
+                    {% if channel.categories | length > 0 %}
+                    {%for cat in channel.categories %}
+                    <tr>
+                        <form method="POST">
+                            <td>{{cat.name}}</td>
+                            <td><input type="submit" value="Löschen" name="deleteCategory"/> <input type="hidden" name="category" value="{{cat.name}}"/></td>
+                        </form>
+                    </tr>
+                    {%endfor%}
+                    {%else%}
+                        <tr><td>keine</td></tr>
+                    {%endif%}
+                </table>
+            </td>
+
+            <td>
+                <form method="POST">
+                    <input list="categories" placeholder="Kategorie" name="category"/>
+                    <datalist id="categories">
+                        {% for cat in categories %}<option value="{{cat.name}}">{% endfor %}
+                    </datalist>
+                    <input type="submit" name="addCategory" value="Hinzufügen"/>
+                </form>
+            </td>
         </tr>
     </table>
 
-
+    <a href="{{url_for('gui.channelView', channel=channel.name)}}">Ansicht</a>
 
     {% else %}
         <h1></h1>
